# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SIPOS - Sistem Informasi Poin Santri API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  admin: 'admin',
  sayyid: 'sayyid',
  wali: 'wali',
} as const;

export interface AuthUser {
  id: number;
  username: string;
  role: AuthUserRole;
  nama: string;
  waliSantriIds?: number[];
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  sayyid: 'sayyid',
  wali: 'wali',
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  nama: string;
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  sayyid: 'sayyid',
  wali: 'wali',
} as const;

export interface UserInput {
  username: string;
  password: string;
  role: UserInputRole;
  nama: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  sayyid: 'sayyid',
  wali: 'wali',
} as const;

export interface UserUpdate {
  username?: string;
  /** @nullable */
  password?: string | null;
  role?: UserUpdateRole;
  nama?: string;
}

export type SantriStatus = typeof SantriStatus[keyof typeof SantriStatus];


export const SantriStatus = {
  aktif: 'aktif',
  nonaktif: 'nonaktif',
  alumni: 'alumni',
} as const;

export interface Santri {
  id: number;
  nis: string;
  nama: string;
  /** @nullable */
  ttl?: string | null;
  /** @nullable */
  jenisKelamin?: string | null;
  kelas: string;
  /** @nullable */
  kamar: string | null;
  asrama: string;
  /** @nullable */
  namaOrtu?: string | null;
  /** @nullable */
  noHp?: string | null;
  status: SantriStatus;
  totalPoin: number;
  createdAt: string;
}

export type SantriDetailStatus = typeof SantriDetailStatus[keyof typeof SantriDetailStatus];


export const SantriDetailStatus = {
  aktif: 'aktif',
  nonaktif: 'nonaktif',
  alumni: 'alumni',
} as const;

export type RiwayatPoinTipe = typeof RiwayatPoinTipe[keyof typeof RiwayatPoinTipe];


export const RiwayatPoinTipe = {
  pelanggaran: 'pelanggaran',
  prestasi: 'prestasi',
} as const;

export interface RiwayatPoin {
  id: number;
  santriId: number;
  /** @nullable */
  santriNama?: string | null;
  tipe: RiwayatPoinTipe;
  /** @nullable */
  masterId?: number | null;
  /** @nullable */
  masterNama?: string | null;
  poin: number;
  /** @nullable */
  keterangan?: string | null;
  tanggal: string;
  /** @nullable */
  createdBy?: string | null;
  createdAt: string;
}

export interface SantriDetail {
  id: number;
  nis: string;
  nama: string;
  /** @nullable */
  ttl?: string | null;
  /** @nullable */
  jenisKelamin?: string | null;
  kelas: string;
  /** @nullable */
  kamar?: string | null;
  asrama: string;
  /** @nullable */
  namaOrtu?: string | null;
  /** @nullable */
  noHp?: string | null;
  status: SantriDetailStatus;
  totalPoin: number;
  createdAt: string;
  riwayatPoin: RiwayatPoin[];
}

export type SantriInputStatus = typeof SantriInputStatus[keyof typeof SantriInputStatus];


export const SantriInputStatus = {
  aktif: 'aktif',
  nonaktif: 'nonaktif',
  alumni: 'alumni',
} as const;

export interface SantriInput {
  nis: string;
  nama: string;
  ttl?: string;
  jenisKelamin?: string;
  kelas: string;
  kamar?: string;
  asrama: string;
  namaOrtu?: string;
  noHp?: string;
  status?: SantriInputStatus;
}

export type SantriUpdateStatus = typeof SantriUpdateStatus[keyof typeof SantriUpdateStatus];


export const SantriUpdateStatus = {
  aktif: 'aktif',
  nonaktif: 'nonaktif',
  alumni: 'alumni',
} as const;

export interface SantriUpdate {
  nis?: string;
  nama?: string;
  /** @nullable */
  ttl?: string | null;
  /** @nullable */
  jenisKelamin?: string | null;
  kelas?: string;
  /** @nullable */
  kamar?: string | null;
  asrama?: string;
  /** @nullable */
  namaOrtu?: string | null;
  /** @nullable */
  noHp?: string | null;
  status?: SantriUpdateStatus;
}

export interface MasterPelanggaran {
  id: number;
  kode: string;
  nama: string;
  poin: number;
  /** @nullable */
  keterangan?: string | null;
  createdAt: string;
}

export interface MasterPelanggaranInput {
  kode: string;
  nama: string;
  poin: number;
  keterangan?: string;
}

export interface MasterPelanggaranUpdate {
  kode?: string;
  nama?: string;
  poin?: number;
  /** @nullable */
  keterangan?: string | null;
}

export interface MasterPrestasi {
  id: number;
  kode: string;
  nama: string;
  poin: number;
  /** @nullable */
  keterangan?: string | null;
  createdAt: string;
}

export interface MasterPrestasiInput {
  kode: string;
  nama: string;
  poin: number;
  keterangan?: string;
}

export interface MasterPrestasiUpdate {
  kode?: string;
  nama?: string;
  poin?: number;
  /** @nullable */
  keterangan?: string | null;
}

export type RiwayatPoinInputTipe = typeof RiwayatPoinInputTipe[keyof typeof RiwayatPoinInputTipe];


export const RiwayatPoinInputTipe = {
  pelanggaran: 'pelanggaran',
  prestasi: 'prestasi',
} as const;

export interface RiwayatPoinInput {
  santriId: number;
  tipe: RiwayatPoinInputTipe;
  masterId: number;
  keterangan?: string;
  tanggal: string;
}

export interface DashboardStats {
  totalSantri: number;
  santriAktif: number;
  pelanggaranHariIni: number;
  prestasiHariIni: number;
  pelanggaranBulanIni: number;
  prestasiBulanIni: number;
}

export interface ActivityLog {
  id: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userName?: string | null;
  aksi: string;
  detail: string;
  createdAt: string;
}

export interface TopSantri {
  id: number;
  nis: string;
  nama: string;
  kelas: string;
  asrama: string;
  totalPoin: number;
  jumlahKejadian: number;
}

export interface GrafikBulanan {
  bulan: string;
  tahun: number;
  pelanggaran: number;
  prestasi: number;
}

export type ListSantriParams = {
search?: string;
kelas?: string;
asrama?: string;
status?: string;
};

export type ListRiwayatPoinParams = {
santriId?: number;
tipe?: ListRiwayatPoinTipe;
tanggalMulai?: string;
tanggalAkhir?: string;
};

export type ListRiwayatPoinTipe = typeof ListRiwayatPoinTipe[keyof typeof ListRiwayatPoinTipe];


export const ListRiwayatPoinTipe = {
  pelanggaran: 'pelanggaran',
  prestasi: 'prestasi',
} as const;

export type GetDashboardTopSantriParams = {
tipe?: GetDashboardTopSantriTipe;
};

export type GetDashboardTopSantriTipe = typeof GetDashboardTopSantriTipe[keyof typeof GetDashboardTopSantriTipe];


export const GetDashboardTopSantriTipe = {
  prestasi: 'prestasi',
  pelanggaran: 'pelanggaran',
} as const;