# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SIPOS - Sistem Informasi Poin Santri API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login pengguna
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Username atau password salah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Mendapatkan data user yang sedang login
      responses:
        "200":
          description: Data user aktif
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Belum login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /master-pelanggaran:
    get:
      operationId: listMasterPelanggaran
      tags: [master-pelanggaran]
      summary: Daftar master pelanggaran
      responses:
        "200":
          description: Daftar master pelanggaran
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MasterPelanggaran"
    post:
      operationId: createMasterPelanggaran
      tags: [master-pelanggaran]
      summary: Tambah master pelanggaran
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterPelanggaranInput"
      responses:
        "201":
          description: Master pelanggaran berhasil ditambahkan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterPelanggaran"
  /master-pelanggaran/{id}:
    put:
      operationId: updateMasterPelanggaran
      tags: [master-pelanggaran]
      summary: Update master pelanggaran
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterPelanggaranUpdate"
      responses:
        "200":
          description: Master pelanggaran berhasil diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterPelanggaran"
    delete:
      operationId: deleteMasterPelanggaran
      tags: [master-pelanggaran]
      summary: Hapus master pelanggaran
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Master pelanggaran berhasil dihapus
  /master-prestasi:
    get:
      operationId: listMasterPrestasi
      tags: [master-prestasi]
      summary: Daftar master prestasi
      responses:
        "200":
          description: Daftar master prestasi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MasterPrestasi"
    post:
      operationId: createMasterPrestasi
      tags: [master-prestasi]
      summary: Tambah master prestasi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterPrestasiInput"
      responses:
        "201":
          description: Master prestasi berhasil ditambahkan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterPrestasi"
  /master-prestasi/{id}:
    put:
      operationId: updateMasterPrestasi
      tags: [master-prestasi]
      summary: Update master prestasi
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterPrestasiUpdate"
      responses:
        "200":
          description: Master prestasi berhasil diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterPrestasi"
    delete:
      operationId: deleteMasterPrestasi
      tags: [master-prestasi]
      summary: Hapus master prestasi
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Master prestasi berhasil dihapus
  /riwayat-poin:
    get:
      operationId: listRiwayatPoin
      tags: [riwayat-poin]
      summary: Daftar riwayat poin
      parameters:
        - name: santriId
          in: query
          schema:
            type: integer
        - name: tipe
          in: query
          schema:
            type: string
            enum: [pelanggaran, prestasi]
        - name: tanggalMulai
          in: query
          schema:
            type: string
        - name: tanggalAkhir
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daftar riwayat poin
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiwayatPoin"
    post:
      operationId: createRiwayatPoin
      tags: [riwayat-poin]
      summary: Tambah riwayat poin (pelanggaran atau prestasi)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiwayatPoinInput"
      responses:
        "201":
          description: Riwayat poin berhasil ditambahkan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiwayatPoin"
  /riwayat-poin/{id}:
    delete:
      operationId: deleteRiwayatPoin
      tags: [riwayat-poin]
      summary: Hapus riwayat poin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Riwayat poin berhasil dihapus
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, role, nama]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, sayyid, wali]
        nama:
          type: string
        waliSantriIds:
          type: array
          items:
            type: integer
...
    MasterPelanggaran:
      type: object
      required: [id, kode, nama, poin, createdAt]
      properties:
        id:
          type: integer
        kode:
          type: string
        nama:
          type: string
        poin:
          type: integer
        keterangan:
          type: ["string", "null"]
        createdAt:
          type: string
    MasterPelanggaranInput:
      type: object
      required: [kode, nama, poin]
      properties:
        kode:
          type: string
        nama:
          type: string
        poin:
          type: integer
        keterangan:
          type: string
    MasterPelanggaranUpdate:
      type: object
      properties:
        kode:
          type: string
        nama:
          type: string
        poin:
          type: integer
        keterangan:
          type: ["string", "null"]
    MasterPrestasi:
      type: object
      required: [id, kode, nama, poin, createdAt]
      properties:
        id:
          type: integer
        kode:
          type: string
        nama:
          type: string
        poin:
          type: integer
        keterangan:
          type: ["string", "null"]
        createdAt:
          type: string
    MasterPrestasiInput:
      type: object
      required: [kode, nama, poin]
      properties:
        kode:
          type: string
        nama:
          type: string
        poin:
          type: integer
        keterangan:
          type: string
    MasterPrestasiUpdate:
      type: object
      properties:
        kode:
          type: string
        nama:
          type: string
        poin:
          type: integer
        keterangan:
          type: ["string", "null"]
    RiwayatPoin:
      type: object
      required: [id, santriId, tipe, poin, tanggal, createdAt]
      properties:
        id:
          type: integer
        santriId:
          type: integer
        santriNama:
          type: ["string", "null"]
        tipe:
          type: string
          enum: [pelanggaran, prestasi]
        masterId:
          type: ["integer", "null"]
        masterNama:
          type: ["string", "null"]
        poin:
          type: integer
        keterangan:
          type: ["string", "null"]
        tanggal:
          type: string
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    RiwayatPoinInput:
      type: object
      required: [santriId, tipe, masterId, tanggal]
      properties:
        santriId:
          type: integer
        tipe:
          type: string
          enum: [pelanggaran, prestasi]
        masterId:
          type: integer
        keterangan:
          type: string
        tanggal:
          type: string
...