# lib/api-zod/src/generated/types/masterPrestasiInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SIPOS - Sistem Informasi Poin Santri API
 * OpenAPI spec version: 0.1.0
 */

export interface MasterPrestasiInput {
  kode: string;
  nama: string;
  poin: number;
  keterangan?: string;
}